Re-organize the folder structure
