+++ src/containers/LoginFormContainer.jsx
@@ -2,16 +2,16 @@ import React from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import LoginForm from './LoginForm';
-import LogoutForm from './LogoutForm';
+import LoginForm from '../components/LoginForm';
+import LogoutForm from '../components/LogoutForm';
 
 import {
   changeLoginField,
   requestLogin,
   logout,
-} from './slice';
+} from '../slice';
 
-import { get } from './utils';
+import { get } from '../utils';
 
 export default function LoginFormContainer() {
   const dispatch = useDispatch();
